ffvl weather stations displayed on map

+++ src/com/geeksville/weather/Stationable.java
@@ -6,7 +6,7 @@ import org.osmdroid.util.GeoPoint;
 
 public interface Stationable {
 	public String getName();
-	public float getAltitude();
+	public int getAltitude();
 	public Measure getMeasure();
 	public GeoPoint getLocation();
 	public Map<String, String> getExtraInfo();
